+++ lib/cocoapods/target/build_settings.rb
@@ -351,6 +351,23 @@ module Pod
 
       #-------------------------------------------------------------------------#
 
+      # @!group Protected Methods
+
+      protected
+
+      # Xcode 11 causes an issue with frameworks or libraries before 12.2 deployment target that link or are part of
+      # test bundles that use XCTUnwrap. Apple has provided an official work around for this.
+      #
+      # @see https://developer.apple.com/documentation/xcode_release_notes/xcode_11_release_notes
+      #
+      # @return [Boolean] Whether to apply the fix or not.
+      #
+      def should_apply_xctunwrap_fix?
+        false
+      end
+
+      #-------------------------------------------------------------------------#
+
       # @!group Private Methods
 
       private
@@ -492,9 +509,15 @@ module Pod
         attr_reader :test_xcconfig
         alias test_xcconfig? test_xcconfig
 
+        # @return [Boolean]
+        #   whether settings are being generated for an application bundle
+        #
         attr_reader :app_xcconfig
         alias app_xcconfig? app_xcconfig
 
+        # @return [Boolean]
+        #   whether settings are being generated for an library bundle
+        #
         attr_reader :library_xcconfig
         alias library_xcconfig? library_xcconfig
 
@@ -683,6 +706,12 @@ module Pod
           file_accessors.flat_map(&:vendored_xcframeworks).map { |path| Xcode::XCFramework.new(path) }
         end
 
+        # @return [Array<String>]
+        define_build_settings_method :system_framework_search_paths, :build_setting => true, :memoized => true, :sorted => true, :uniqued => true do
+          return ['$(PLATFORM_DIR)/Developer/usr/lib'] if should_apply_xctunwrap_fix?
+          []
+        end
+
         #-------------------------------------------------------------------------#
 
         # @!group Libraries
@@ -737,16 +766,18 @@ module Pod
 
         # @return [Array<String>]
         define_build_settings_method :library_search_paths, :build_setting => true, :memoized => true, :sorted => true, :uniqued => true do
-          return [] if library_xcconfig? && target.build_as_static?
+          library_search_paths = should_apply_xctunwrap_fix? ? ['$(PLATFORM_DIR)/Developer/usr/lib'] : []
+          return library_search_paths if library_xcconfig? && target.build_as_static?
 
-          vendored = library_search_paths_to_import.dup
-          vendored.concat dependent_targets.flat_map { |pt| pt.build_settings[@configuration].vendored_dynamic_library_search_paths }
+          library_search_paths.concat library_search_paths_to_import.dup
+          library_search_paths.concat dependent_targets.flat_map { |pt| pt.build_settings[@configuration].vendored_dynamic_library_search_paths }
           if library_xcconfig?
-            vendored.delete(target.configuration_build_dir(CONFIGURATION_BUILD_DIR_VARIABLE))
+            library_search_paths.delete(target.configuration_build_dir(CONFIGURATION_BUILD_DIR_VARIABLE))
           else
-            vendored.concat(dependent_targets.flat_map { |pt| pt.build_settings[@configuration].library_search_paths_to_import })
+            library_search_paths.concat(dependent_targets.flat_map { |pt| pt.build_settings[@configuration].library_search_paths_to_import })
           end
-          vendored
+
+          library_search_paths
         end
 
         # @return [Array<String>]
@@ -837,6 +868,7 @@ module Pod
         define_build_settings_method :swift_include_paths, :build_setting => true, :memoized => true, :sorted => true, :uniqued => true do
           paths = dependent_targets.flat_map { |pt| pt.build_settings[@configuration].swift_include_paths_to_import }
           paths.concat swift_include_paths_to_import if non_library_xcconfig?
+          paths.concat ['$(PLATFORM_DIR)/Developer/usr/lib'] if should_apply_xctunwrap_fix?
           paths
         end
 
@@ -966,6 +998,12 @@ module Pod
           end
         end
 
+        # @see BuildSettings#should_apply_xctunwrap_fix?
+        def should_apply_xctunwrap_fix?
+          library_xcconfig? && Version.new(target.platform.deployment_target) < Version.new('12.2') &&
+            consumer_frameworks.include?('XCTest')
+        end
+
         #-------------------------------------------------------------------------#
       end
 
@@ -1037,6 +1075,12 @@ module Pod
           []
         end
 
+        # @return [Array<String>]
+        define_build_settings_method :system_framework_search_paths, :build_setting => true, :memoized => true, :sorted => true, :uniqued => true do
+          return ['$(PLATFORM_DIR)/Developer/usr/lib'] if should_apply_xctunwrap_fix?
+          []
+        end
+
         #-------------------------------------------------------------------------#
 
         # @!group Libraries
@@ -1048,6 +1092,7 @@ module Pod
 
         # @return [Array<String>]
         define_build_settings_method :library_search_paths, :build_setting => true, :memoized => true, :sorted => true, :uniqued => true, :from_pod_targets_to_link => true, :from_search_paths_aggregate_targets => :vendored_dynamic_library_search_paths do
+          return ['$(PLATFORM_DIR)/Developer/usr/lib'] if should_apply_xctunwrap_fix?
           []
         end
 
@@ -1113,6 +1158,7 @@ module Pod
 
         # @return [Array<String>]
         define_build_settings_method :swift_include_paths, :build_setting => true, :memoized => true, :sorted => true, :uniqued => true, :from_pod_targets_to_link => true, :from_search_paths_aggregate_targets => :swift_include_paths_to_import do
+          return ['$(PLATFORM_DIR)/Developer/usr/lib'] if should_apply_xctunwrap_fix?
           []
         end
 
@@ -1244,6 +1290,17 @@ module Pod
         #
         define_build_settings_method :merged_user_target_xcconfigs, :memoized => true do
           merged_xcconfigs(user_target_xcconfig_values_by_consumer_by_key, :user_target_xcconfig)
+        end
+
+        # @see BuildSettings#should_apply_xctunwrap_fix?
+        def should_apply_xctunwrap_fix?
+          target.library? && target.user_targets.any? do |ut|
+            next false unless Version.new(ut.deployment_target) < Version.new('12.2')
+            flags = (ut.resolved_build_setting('OTHER_LDFLAGS', :resolve_against_xcconfig => true) || {}).values.flatten.uniq
+            parsed_flags = Xcodeproj::Config::OtherLinkerFlagsParser.parse(flags)
+            all_frameworks = (parsed_flags[:frameworks] + parsed_flags[:weak_frameworks]).uniq
+            all_frameworks.include?('XCTest') || all_frameworks.include?('"XCTest"')
+          end
         end
 
         #-------------------------------------------------------------------------#